northd: find outports based on ovn_port
Previously we searched for the outport of a route by going through all
LRPs of the LR in the northbound, finding the matching ovn_port and
checking if it matches. However we can also just iterate over all
ovn_port of the LR datapath.

This is simpler and a prerequisite for later patches that use
derived router ports.