DO NOT APPLY: Use my ovs repo and bump.
this patchset depends on some additional features in the OVS repo. As
they are not yet in there at the time of development i just reference to
my repo here to make ci testing work.

Ideally i have sent a link in the cover-letter to the OVS patchset that
is needed for this change. Once this is merged we should replace this
commit with a bump of the ovs repo.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>