controller: Allow filtering routes on ifindex.
previously if a single chassis has multiple LRPs of a LR bound then the
routes learned would be shared by all of these LRPs. However potentially
only some of the LRPs have actually received this route.

To allow for filtering we allow the user to define the ifindex of each
interface and thereby learn routes individually per LRP.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>