tests: Fix flanky I+P tests.
The I+P test for dynamic-routing previously specified if engine nodes
need to have recomputed and/or computed for a specific change.
However based on the execution speed of the test some changes may be
batched together in northd (especially northbound and southbound
updates).

For nodes where we specified "recompute compute" we sometimes did not
get a nonzero "compute" value as the changes where handled in a run that
also recomputed.

However this test should only find out if specific changes are processed
fully incrementally. As such we only care about 3 potential options:
* We did a recompute (so no or partial incremental handling)
* We did no recompute and handled a change incrementally
* There where no changes

By changing to this test method we avoid the issue above.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>