tests: Handle proto ovn in iproute.
The rt_proto 84 has been claimed upstream for OVN. This mapping is now
also added to iproute2 [1].
Once a version with this commit is in use "ip route list" will show
"proto ovn" instead of "proto 84". Since this depends on the iproute2
version we handle both cases in the tests.

[1]: https://git.kernel.org/pub/scm/network/iproute2/iproute2-next.git/commit/?id=131810fb264f

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>