route: Only handle datapaths that should advertise.
Previously we created an advertise_datapath_entry for each local router
datapath, independently if any of its ports are bound to the current
chassis or take part in dynamic-routing at all.

As ad->maintain_vrf was false and there was never a route added this did
not do anything other than wasting memory.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>