controller: Fix advertised_route incremental handler.
There can be endge cases where an advertised_route has a tracked_port
field but the local IDL has not yet loaded the corresponding logical
port.
Once that issue resolves we need to trigger a recompute of the route
node so that tracked_ports_local and tracked_ports_remote can be fixed.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>